requirements.txt
langchain==1.2.7
langchain-deepseek==1.0.1
python-dotenv>=1.0.0
fastapi>=0.110
uvicorn[standard]>=0.24
